--- src/main/java/dandi/dandi/auth/application/OAuthClient.java
@@ -1,4 +1,0 @@
-package dandi.dandi.auth.application;
-
-public interface OAuthClient {
-}
