+++ src/main/java/dandi/dandi/auth/domain/OAuthClient.java
@@ -1,0 +1,4 @@
+package dandi.dandi.auth.domain;
+
+public interface OAuthClient {
+}
