+++ src/main/java/dandi/dandi/auth/infrastructure/apple/dto/ApplePublicKeys.java
@@ -1,4 +1,4 @@
-package dandi.dandi.auth.application.api;
+package dandi.dandi.auth.infrastructure.apple.dto;
 
 import java.util.List;
